Extract filter building into FilterBuilder
